adicionando la funcionalidad para mas de una imagen en el crear y en el editar vestido en el admin, y haciendo navegable las imagenes en el cliente

+++ src/routes/api/size/[id]/+server.ts
@@ -1,9 +1,10 @@
+import z from 'zod';
 import { eq } from 'drizzle-orm';
 import { json } from '@sveltejs/kit';
 import { db } from '$lib/server/index.js';
 import { size } from '$lib/server/db/schema';
 import type { RequestHandler } from './$types';
-import type { SizeDto } from '$lib/server/types/Dto';
+import { sizeUpdateSchema } from '$lib/server/types/models';
 
 
 export const GET: RequestHandler = async ({ params }) => {
@@ -14,9 +15,7 @@ export const GET: RequestHandler = async ({ params }) => {
 			return json({ message: 'invalid ID '}, { status: 400 });
 		}
 
-		const result = await db.query.size.findFirst({
-			where: eq(size.id, id)
-		});
+		const result = await db.query.size.findFirst({ where: eq(size.id, id) });
 
 		if (!result) {
 			return json({ message: 'Size not found'}, { status: 404 });
@@ -43,10 +42,23 @@ export const PUT: RequestHandler = async ({ request, params, cookies }) => {
 			return json({ message: 'invalid ID '}, { status: 400 });
 		}
 
-		const body: Partial<SizeDto> = await request.json();
+		const response = await request.json();
+		const data_validated = await sizeUpdateSchema.safeParseAsync(response);
+
+		if(!data_validated.success){
+			return json({
+				errors: z.flattenError(data_validated.error).fieldErrors
+			});
+		}
+
+		const existing = await db.select().from(size).where(eq(size.id, id)).limit(1);
+
+		if (existing.length === 0) {
+			return json({ message: 'Size not found' }, { status: 404 });
+		}
 
 		const update = await db.update(size)
-								.set(body)
+								.set(response)
 								.where( eq(size.id, id) )
 								.returning({ id: size.id, size: size.size, height: size.height });
 		
@@ -73,6 +85,12 @@ export const DELETE: RequestHandler = async ({ params, cookies }) => {
 
 		if (isNaN(id)) {
 			return json({ message: 'invalid ID '}, { status: 400 });
+		}
+
+		const existing = await db.select().from(size).where(eq(size.id, id)).limit(1);
+
+		if (existing.length === 0) {
+			return json({ message: 'Size not found' }, { status: 404 });
 		}
 
 		const result = await db.delete(size)
